% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolmato.R
\name{install_n_load}
\alias{install_n_load}
\title{Load required packages. Install them if not installed already.}
\usage{
install_n_load(.required_packages)
}
\arguments{
\item{.required_packages}{A character list of the packages you need to load.}
}
\value{

}
\description{
Load required packages. Install them if not installed already.
}
\examples{
Required_Packages=c("data.table", "lubridate","DT","tidyr", "dplyr")

install_n_load(Required_Packages)
}
